+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+**Features**:
+
+- Add new `relay-threading` crate with asynchronous thread pool. ([#4500](https://github.com/getsentry/relay/pull/4500))
+
+## 25.2.0
+
 - Allow log ingestion behind a flag, only for internal use currently. ([#4471](https://github.com/getsentry/relay/pull/4471))
 
 **Features**:
@@ -26,6 +32,7 @@
 - Add option to drop transaction attachments. ([#4466](https://github.com/getsentry/relay/pull/4466))
 - Add endpoint that exposes internally collected relay metrics. ([#4497](https://github.com/getsentry/relay/pull/4497))
 - Add sub-millisecond precision to internal timer metrics. ([#4495](https://github.com/getsentry/relay/pull/4495))
+- Partition spans by `trace_id` on the Kafka topic. ([#4503](https://github.com/getsentry/relay/pull/4503))
 
 ## 25.1.0